requirements.txt
opencv_python==4.5.3.56
Flask==2.0.1
numpy==1.21.2
mediapipe==0.8.7.3
Pillow==8.3.2
